requirements.txt
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
autopep8==1.4
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
botocore==1.12.50
Bottleneck==1.2.1
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.2.2
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
distributed==1.21.8
Django==1.10.8
django-classy-tags==0.8.0
django-sekizai==0.10.0
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
facebook-sdk==3.1.0
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
foundation==0.1.0a0.dev1
gevent==1.3.0
glob2==0.6
gmpy2==2.0.8
googletrans==2.3.0
greenlet==0.4.13
gTTS==2.0.1
gTTS-token==1.1.1
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kiwisolver==1.0.1
langdetect==1.0.7
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.4.5
psycopg2==2.7.6.1
ptyprocess==0.5.2
py==1.5.3
py-translator==1.8.9
pycairo==1.17.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
pyobjc==4.2.2
pyobjc-core==4.2.2
pyobjc-framework-Accounts==4.2.2
pyobjc-framework-AddressBook==4.2.2
pyobjc-framework-AppleScriptKit==4.2.2
pyobjc-framework-AppleScriptObjC==4.2.2
pyobjc-framework-ApplicationServices==4.2.2
pyobjc-framework-Automator==4.2.2
pyobjc-framework-AVFoundation==4.2.2
pyobjc-framework-AVKit==4.2.2
pyobjc-framework-BusinessChat==4.2.2
pyobjc-framework-CalendarStore==4.2.2
pyobjc-framework-CFNetwork==4.2.2
pyobjc-framework-CloudKit==4.2.2
pyobjc-framework-Cocoa==4.2.2
pyobjc-framework-Collaboration==4.2.2
pyobjc-framework-ColorSync==4.2.2
pyobjc-framework-Contacts==4.2.2
pyobjc-framework-ContactsUI==4.2.2
pyobjc-framework-CoreBluetooth==4.2.2
pyobjc-framework-CoreData==4.2.2
pyobjc-framework-CoreLocation==4.2.2
pyobjc-framework-CoreML==4.2.2
pyobjc-framework-CoreServices==4.2.2
pyobjc-framework-CoreSpotlight==4.2.2
pyobjc-framework-CoreText==4.2.2
pyobjc-framework-CoreWLAN==4.2.2
pyobjc-framework-CryptoTokenKit==4.2.2
pyobjc-framework-DictionaryServices==4.2.2
pyobjc-framework-DiskArbitration==4.2.2
pyobjc-framework-EventKit==4.2.2
pyobjc-framework-ExceptionHandling==4.2.2
pyobjc-framework-ExternalAccessory==4.2.2
pyobjc-framework-FinderSync==4.2.2
pyobjc-framework-FSEvents==4.2.2
pyobjc-framework-GameCenter==4.2.2
pyobjc-framework-GameController==4.2.2
pyobjc-framework-GameKit==4.2.2
pyobjc-framework-GameplayKit==4.2.2
pyobjc-framework-ImageCaptureCore==4.2.2
pyobjc-framework-IMServicePlugIn==4.2.2
pyobjc-framework-InputMethodKit==4.2.2
pyobjc-framework-InstallerPlugins==4.2.2
pyobjc-framework-InstantMessage==4.2.2
pyobjc-framework-Intents==4.2.2
pyobjc-framework-IOSurface==4.2.2
pyobjc-framework-iTunesLibrary==4.2.2
pyobjc-framework-LatentSemanticMapping==4.2.2
pyobjc-framework-LaunchServices==4.2.2
pyobjc-framework-libdispatch==4.2.2
pyobjc-framework-LocalAuthentication==4.2.2
pyobjc-framework-MapKit==4.2.2
pyobjc-framework-MediaAccessibility==4.2.2
pyobjc-framework-MediaLibrary==4.2.2
pyobjc-framework-MediaPlayer==4.2.2
pyobjc-framework-ModelIO==4.2.2
pyobjc-framework-MultipeerConnectivity==4.2.2
pyobjc-framework-NetFS==4.2.2
pyobjc-framework-NetworkExtension==4.2.2
pyobjc-framework-NotificationCenter==4.2.2
pyobjc-framework-OpenDirectory==4.2.2
pyobjc-framework-Photos==4.2.2
pyobjc-framework-PhotosUI==4.2.2
pyobjc-framework-PreferencePanes==4.2.2
pyobjc-framework-PubSub==4.2.2
pyobjc-framework-QTKit==4.2.2
pyobjc-framework-Quartz==4.2.2
pyobjc-framework-SafariServices==4.2.2
pyobjc-framework-SceneKit==4.2.2
pyobjc-framework-ScreenSaver==4.2.2
pyobjc-framework-ScriptingBridge==4.2.2
pyobjc-framework-SearchKit==4.2.2
pyobjc-framework-Security==4.2.2
pyobjc-framework-SecurityFoundation==4.2.2
pyobjc-framework-SecurityInterface==4.2.2
pyobjc-framework-ServiceManagement==4.2.2
pyobjc-framework-Social==4.2.2
pyobjc-framework-SpriteKit==4.2.2
pyobjc-framework-StoreKit==4.2.2
pyobjc-framework-SyncServices==4.2.2
pyobjc-framework-SystemConfiguration==4.2.2
pyobjc-framework-Vision==4.2.2
pyobjc-framework-WebKit==4.2.2
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-espeak==0.1.0
python-slugify==1.2.6
pyttsx3==2.7
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.18.4
requests-oauthlib==1.0.0
rope==0.10.7
ruamel-yaml==0.15.35
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
treetaggerwrapper==2.2.4
tweepy==3.6.0
typing==3.6.4
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
vk==2.0.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.2.0
zict==0.1.3
